Capstone Project
Problem Statement
Based on the laboratory tests collected from the suspected cases, predict the chances of being positive or negative for covid19 and identify the factors that influence it. Also, provide the recommendations to the hospital on how they can better manage the admission of patients to the general ward, semi-intensive unit, or intensive care unit.

Need of the study/project
One of the motivations for this problem was the fact that in the context of an overwhelmed health system with the possible limitation to performing tests for the detection of SARS-CoV-2, testing every case with some mild symptoms like cold & cough or suffering from mild fever would be impractical and test results could be delayed because of more number of people complaining about mild symptoms & going for tests. So setting up criteria to get tested was very important to focus on getting the test results early & also to reduce the spread of the virus quicker.

To slow the virus's spread across the country & to minimize the number of hospitalizations with severity.

Understanding business/social opportunity
How the hospital can better manage the admission of patients to the general ward, semi-intensive unit, or intensive care unit. The decease spreads so rapidly and large number of cases are coming everyday but the number of hospital beds are limited. Its needed to study deeply to save maximum number of life.